Change float to double

+++ Client/Scripts/Gameplay/Abilities/BloodSacrificeAbility.cs
@@ -9,7 +9,7 @@ public sealed class BloodSacrificeAbility : Ability
     public override CanTarget Targeting => CanTarget.Nothing;
     public override Scalable<double> StaticHealthCost => 100;
     public override Scalable<double> StaticUseTime => 0.5;
-    private Scalable<float> ManaRestored => 200;
+    private Scalable<double> ManaRestored => 200;
     public override Scalable<double> StaticCooldown => 5;
 
     public override void TakeEffect(AbilityContext context)
