+++ Client/Scripts/Gameplay/Interfaces/IEntityManager.cs
@@ -14,7 +14,7 @@ public interface IEntityManager
     
     void ReplicateSnapshotEntities(ShardSnapshot snapshot);
     PlayerCharacter SpawnPlayerCharacter(Guid id);
-    AttackProjectile SpawnAttackProjectile(AbilityContext abilityContext, float speed);
+    AttackProjectile SpawnAttackProjectile(AbilityContext abilityContext, double speed);
     T? GetEntity<T>(Guid id);
     IEntity? GetEntity(Guid id);
 }
